backend/requirements.txt
FastApi
pydantic
uvicorn
pre-commit
flake8
black
Adafruit-Blinka
adafruit-circuitpython-neopixel
aioredis
Pillow